// src/interfaces/price-table.interface.ts
export interface PriceTableInterface {
  brand: string;
  model: string;
  year: string;
}

export interface PriceTableResultInterface {
  Marca: string;
  Valor: string;
  Modelo: string;
  AnoModelo: number;
  CodigoFipe: string;
  Combustivel: string;
  TipoVeiculo: number;
  MesReferencia: string;
  SiglaCombustivel: string;
}